check robust.f

+++ src/SearchLibrium/boxcox_functions.py
@@ -55,14 +55,14 @@ def transform(x, lmda):
 def transform_derivative(x, lmda):
 # {
     """ Derivative of log likelihood with respect to lmda """
-    ln_x = np.log1p(x) # Evaluate ln(1+x) because elements of x are close to zero
+    ln_x = np.log1p(x.astype(float)) # Evaluate ln(1+x) because elements of x are close to zero
     # Note: original code was ln_x = np.log(x)
 
     if lmda == 0:
         return 0.5*(ln_x)**2  # i.e., 0.5 * ln(x)^2
     else:
     # {
-        x_lmda = np.nan_to_num(np.power(x, lmda))
+        x_lmda = np.nan_to_num(np.power(x.astype(float), lmda))
         val = (lmda * x_lmda * ln_x - x_lmda + 1) / np.power(lmda, 2)
         return np.nan_to_num(val) # Return zero
     # }
